driver:src:Modify XAieEventNum to support u16
This patch contain below changes
1.XAIE_EVENT_INVALID changed to 256
2.changed XAieEventNum from u8 type to u16.

+++ driver/src/global/xaie2ipugbl_reginit.c
@@ -2422,7 +2422,7 @@ static const  XAie_LockMod AieMlMemTileLockMod =
 
 #ifdef XAIE_FEATURE_EVENTS_ENABLE
 /* Enum to event number mapping of all events of AIEML Core Mod of aie tile */
-static const u8 AieMlCoreModEventMapping[] =
+static const u16 AieMlCoreModEventMapping[] =
 {
 	XAIEML_EVENTS_CORE_NONE,
 	XAIEML_EVENTS_CORE_TRUE,
@@ -2560,7 +2560,7 @@ static const u8 AieMlCoreModEventMapping[] =
 };
 
 /* Enum to event number mapping of all events of AIEML Mem Mod of aie tile */
-static const u8 AieMlMemModEventMapping[] =
+static const u16 AieMlMemModEventMapping[] =
 {
 	XAIEML_EVENTS_MEM_NONE,
 	XAIEML_EVENTS_MEM_TRUE,
@@ -2727,7 +2727,7 @@ static const u8 AieMlMemModEventMapping[] =
 };
 
 /* Enum to event number mapping of all events of AIEML NOC tile */
-static const u8 AieMlNocModEventMapping[] =
+static const u16 AieMlNocModEventMapping[] =
 {
 	XAIEML_EVENTS_PL_NONE,
 	XAIEML_EVENTS_PL_TRUE,
@@ -2905,7 +2905,7 @@ static const u8 AieMlNocModEventMapping[] =
 };
 
 /* Enum to event number mapping of all events of AIEML Mem Tile Module */
-static const u8 AieMlMemTileModEventMapping[] =
+static const u16 AieMlMemTileModEventMapping[] =
 {
 	XAIEML_EVENTS_MEM_TILE_NONE,
 	XAIEML_EVENTS_MEM_TILE_TRUE,
