+++ driver/src/global/xaiegbl_regdef.h
@@ -746,7 +746,7 @@ typedef struct {
 
 /* This typedef contains attributes of Events module */
 typedef struct XAie_EvntMod {
-	const u8 *XAie_EventNumber;	/* Array of event numbers with true event val */
+	const u16 *XAie_EventNumber;	/* Array of event numbers with true event val */
 	u8  NumEventReg;
 	u32 EventMin;		/* number corresponding to evt 0 in the enum */
 	u32 EventMax;		/* number corresponding to last evt in enum */
